Refactor table search into TableSearchBar component with query prop
